+++ src/full/Agda/Interaction/Options/Base.hs
@@ -79,6 +79,7 @@ module Agda.Interaction.Options.Base
     , lensOptErasure
     , lensOptErasedMatches
     , lensOptEraseRecordParameters
+    , lensOptErasedLevelsInPrims
     , lensOptRewriting
     , lensOptLocalRewriting
     , lensOptCubical
@@ -145,6 +146,7 @@ module Agda.Interaction.Options.Base
     , optErasure
     , optErasedMatches
     , optEraseRecordParameters
+    , optErasedLevelsInPrims
     , optRewriting
     , optLocalRewriting
     , optGuarded
@@ -315,6 +317,7 @@ optProjectionLike            :: PragmaOptions -> Bool
 optErasure                   :: PragmaOptions -> Bool
 optErasedMatches             :: PragmaOptions -> Bool
 optEraseRecordParameters     :: PragmaOptions -> Bool
+optErasedLevelsInPrims       :: PragmaOptions -> Bool
 optRewriting                 :: PragmaOptions -> Bool
 optLocalRewriting            :: PragmaOptions -> Bool
 optGuarded                   :: PragmaOptions -> Bool
@@ -380,6 +383,7 @@ optProjectionLike            = collapseDefault . _optProjectionLike
 optErasure                   = collapseDefault . _optErasure || optEraseRecordParameters || (Value True ==) . _optErasedMatches
 optErasedMatches             = collapseDefault . _optErasedMatches && optErasure
 optEraseRecordParameters     = collapseDefault . _optEraseRecordParameters
+optErasedLevelsInPrims       = collapseDefault . _optErasedLevelsInPrims
 optRewriting                 = collapseDefault . _optRewriting
 optLocalRewriting            = collapseDefault . _optLocalRewriting
 optGuarded                   = collapseDefault . _optGuarded
@@ -550,6 +554,11 @@ lensOptErasedMatches f o = f (_optErasedMatches o) <&> \ i -> o{ _optErasedMatch
 
 lensOptEraseRecordParameters :: Lens' PragmaOptions _
 lensOptEraseRecordParameters f o = f (_optEraseRecordParameters o) <&> \ i -> o{ _optEraseRecordParameters = i }
+
+lensOptErasedLevelsInPrims :: Lens' PragmaOptions _
+lensOptErasedLevelsInPrims f o =
+  f (_optErasedLevelsInPrims o) <&>
+  \i -> o { _optErasedLevelsInPrims = i }
 
 lensOptRewriting :: Lens' PragmaOptions _
 lensOptRewriting f o = f (_optRewriting o) <&> \ i -> o{ _optRewriting = i }
@@ -952,6 +961,10 @@ infectiveCoinfectiveOptions =
   , infectiveOption optCohesion               "--cohesion"
   , infectiveOption optErasure                "--erasure"
   , infectiveOption optErasedMatches          "--erased-matches"
+  , infectiveOption (not . optErasedLevelsInPrims)
+      "--no-erased-levels-in-primitives"
+  , coinfectiveOption (not . optErasedLevelsInPrims)
+      "--no-erased-levels-in-primitives"
   ]
   where
   cubicalCompatible =
@@ -1560,6 +1573,11 @@ modalityPragmaOptions = ("Modalities",) $ concat
   , pragmaFlag      "erase-record-parameters" lensOptEraseRecordParameters
                     "mark all parameters of record modules as erased" "(implies --erasure)"
                     Nothing
+  , [ Option []     ["no-erased-levels-in-primitives"]
+                    (NoArg $ \o ->
+                     return o { _optErasedLevelsInPrims = Value False })
+                    "do not use erased levels in primitives/builtins"
+    ]
   , pragmaFlag      "cohesion" lensOptCohesion
                     "enable the cohesion modalities" "(in particular @flat)"
                     Nothing