+++ src/full/Agda/TypeChecking/Rules/Data.hs
@@ -504,7 +504,7 @@ defineCompData d con params names fsT t boundary = do
               imax x y = pure tIMax <@> x <@> y
               ineg r = pure tINeg <@> r
               lvlOfType = (\ (Type l) -> Level l) . getSort
-              pOr la i j u0 u1 = pure tPOr <#> (lvlOfType <$> la) <@> i <@> j
+              pOr la i j u0 u1 = pure tPOr <#@> (lvlOfType <$> la) <@> i <@> j
                                            <#> ilam "o" (\ _ -> unEl <$> la) <@> u0 <@> u1
               absAp x y = liftM2 absApp x y
 
@@ -523,7 +523,7 @@ defineCompData d con params names fsT t boundary = do
                   let
                     -- Γ, i ⊢ squeeze u = primTrans (\ j -> ty [i := i ∨ j]) (φ ∨ i) u
                     squeeze u = cl primTrans
-                                          <#> lam "j" (\ j -> lvlOfType <$> ty `absAp` (imax i j))
+                                          <#@> lam "j" (\ j -> lvlOfType <$> ty `absAp` (imax i j))
                                           <@> lam "j" (\ j -> unEl <$> ty `absAp` (imax i j))
                                           <@> (phi `imax` i)
                                           <@> u
@@ -546,7 +546,7 @@ defineCompData d con params names fsT t boundary = do
                 faces <- mapM (\ x -> liftM2 (,) (open . noabsApp __IMPOSSIBLE__ $ fmap fst x) (open $ fmap snd x)) faces
                 let
                   thePsi = foldl1 imax (map fst faces)
-                  hcomp ty phi sys a0 = pure tHComp <#> (lvlOfType <$> ty)
+                  hcomp ty phi sys a0 = pure tHComp <#@> (lvlOfType <$> ty)
                                                     <#> (unEl <$> ty)
                                                     <#> phi
                                                     <@> sys
@@ -968,7 +968,7 @@ defineConClause trD' isHIT mtrX npars nixs xTel' telI sigma dT' cnames = do
           ty <- open $ ty
           face <- (foldr max (pure iz) $ map fst $ sys)
           sys <- lam "i'" $ \ i -> combineSys l ty [(phi, u <@> i) | (phi,u) <- sys]
-          pure tHComp <#> l <#> ty <#> pure face <@> pure sys <@> u0
+          pure tHComp <#@> l <#> ty <#> pure face <@> pure sys <@> u0
   interval <- primIntervalType
   let intervalTel nm = ExtendTel (defaultDom interval) (Abs nm EmptyTel)
 
@@ -1023,7 +1023,9 @@ defineConClause trD' isHIT mtrX npars nixs xTel' telI sigma dT' cnames = do
         let
           origPHComp = do
             LEl l t <- fromMaybe __IMPOSSIBLE__ <.> toLType =<< (dT `applyN` (delta ++ x ++ [pure iz]))
-            let ds = map (argH . unnamed . dotP) [Level l, t]
+            let f = unnamed . dotP
+            flArg <- argE (f (Level l))
+            let ds = [flArg, argH (f t)]
             sequence as0 >>= \case
               ps0@[_hphi,_u,_u0] ->
                 pure $ DefP defaultPatternInfo qHComp $ ds ++ ps0
@@ -1453,13 +1455,13 @@ defineTranspForFields pathCons applyProj name params fsT fns rect = do
         case unDom filled_ty' of
           LType (LEl l _) -> do
             let lvl = lam_i $ Level l
-            return $ runNames [] $ do
+            runNamesT [] $ do
              lvl <- open lvl
              phi <- open the_phi
              field <- open field
-             pure transp <#> lvl <@> pure filled_ty
-                                 <@> phi
-                                 <@> field
+             pure transp <#@> lvl <@> pure filled_ty
+                                  <@> phi
+                                  <@> field
           -- interval arg
           ClosedType{}  ->
             return $ runNames [] $ do
@@ -1523,6 +1525,7 @@ defineHCompForFields applyProj name params fsT fns rect = do
   transp <- getPrimitiveTerm builtinTrans
   por <- getPrimitiveTerm builtinPOr
   one <- primItIsOne
+  (<#@>) <- mkEApp
   reportSDoc "comp.rec" 20 $ text $ show params
   reportSDoc "comp.rec" 20 $ text $ show delta
   reportSDoc "comp.rec" 10 $ text $ show fsT
@@ -1585,7 +1588,7 @@ defineHCompForFields applyProj name params fsT fns rect = do
         lam "i" $ \ i -> do
           args <- sequence params
           psi  <- pure imax <@> phi <@> (pure ineg <@> i)
-          u <- lam "j" (\ j -> pure por <#> lvl
+          u <- lam "j" (\ j -> pure por <#@> lvl
                                         <@> phi
                                         <@> (pure ineg <@> i)
                                         <#> lam "_" (\ o -> rect)
@@ -1612,12 +1615,12 @@ defineHCompForFields applyProj name params fsT fns rect = do
   comp <- do
         let
           imax i j = pure tIMax <@> i <@> j
-        let forward la bA r u = pure transp <#> lam "i" (\ i -> la <@> (i `imax` r))
+        let forward la bA r u = pure transp <#@> lam "i" (\ i -> la <@> (i `imax` r))
                                             <@> lam "i" (\ i -> bA <@> (i `imax` r))
                                             <@> r
                                             <@> u
         return $ \ la bA phi u u0 ->
-          pure hcomp <#> (la <@> pure io) <#> (bA <@> pure io) <#> phi
+          pure hcomp <#@> (la <@> pure io) <#> (bA <@> pure io) <#> phi
                       <@> lam "i" (\ i -> ilam "o" $ \ o ->
                               forward la bA i (u <@> i <..> o))
                       <@> forward la bA (pure iz) u0